pyproject.toml
[project]
name = "green-signal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "google-genai>=1.46.0",
]
